pyproject.toml
[project]
name = "splatoon3-wiki-pipeline"
version = "0.1.0"
description = "LLM Wiki generation pipeline for splatoon3-wiki"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "google-genai>=0.3.0",
    "pydantic>=2.8.0",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "RUF"]
# Japanese content uses characters like "×" (U+00D7) and "ー" that ruff flags
# as ambiguous. Disable RUF001/002/003 project-wide.
ignore = ["RUF001", "RUF002", "RUF003", "UP047"]
